Отрефакторит презентер и модуль с утилитами
