+++ src/const.js
@@ -1,4 +1,4 @@
-const WAYPPOINTS_COUNT = 4;
+const WAYPPOINTS_COUNT = 10;
 const DEFAULT_DATE_FORMAT = 'YYYY-MM-DDTHH:mm:ss.sssZ';
 const SHORT_DEFAULT_FORMAT = 'YYYY-MM-DD';
 const SHORT_DATE_FORMAT = 'MMM D';
